fix: documentation and minor fixes

+++ src/lexer/lexer-utils.c
@@ -6,7 +6,7 @@
 /*   By: ppontet <ppontet@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/11 11:01:21 by ppontet           #+#    #+#             */
-/*   Updated: 2025/04/11 11:08:19 by ppontet          ###   ########lyon.fr   */
+/*   Updated: 2025/04/15 11:50:47 by ppontet          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,17 +20,24 @@ char	*ft_substr_end(char const *src, unsigned int start, size_t end);
 /**
  * @brief Is |, < or >
  *
- * @param c
- * @return int
+ * @param c character to check
+ * @return int 1 if c is operator, 0 otherwise
  */
 int	is_operator_char(char c)
 {
 	return (c == '|' || c == '<' || c == '>');
 }
-
+/**
+ * @brief Allocate memory for a string and copy the content from src
+ *
+ * @param src source string
+ * @param start start index to copy from
+ * @param end end index to copy to
+ * @return char* pointer to the new string
+ */
 char	*ft_substr_end(char const *src, unsigned int start, size_t end)
 {
-	int		len;
+	size_t		len;
 	char	*s;
 
 	len = end - start;
